Extract the transferable state into a separate cls
The visual state needs to be transferrable between TE navigations
since the view and vm are disposed each time.
I've also simplified the tests, the API and removed the bound-projects (plural)
concept and replaced it with a single HasBoundProject property which will work
with WPF as long as the state API will be used to set/clear the bound project.