Fix issue - removing Includes under path was wrong
Corrected an issue whereby when updating a project level ruleset's
includes, the removal of old 'SonarQube' rule sets wasn't happening.
In fact it was (in some cases) removing other rule set includes NOT under
the SonarQube folder.

This was because the relative path resolution of the rule set include was
being done against the wrong base path.