Alleviate file locking issues by reducing no. writes
For solutions which already had setup shared project-level rule sets, we
were updating the rule set file once per project (with the same include).
This was causing unnessesary disk writes. As a result the IDE/Roslyn was
reloading (with a lock for some reason) this file just as we are about to
write to it again (for the next project).

Now keeping track of which existing rule sets we've already updated and
skipping the update & write when we try with the same file again.
This is OK since a new ProjectRuleSetWriter is created for each
BindingWorkflow (which is per bind operation).