Prepare for solution level conficts detection check once synced
1. Moved all service-liked interface to be serviced by the host - this helped to simplify the test specific service injection and simplifyed the product code that in most cases no longer needs those interfaces to be passed from the constructor.

2. Extracted the logic from binding into a service that will be used in the last piece (i.e. find and fix conflicts in the entire solution).

I also included a code example (BindingController.CheckForConflicts) that I'll be aiming for in the last PR for this feature. In general there will be a service that I can invoke from other places and not just after binding, although in those case we will need to address some of the edge cases which I don't plan on handling at first (need to get initial feedback from Jean-Marc first).