Integrated the previously reviewed components into the extension
1. Added ConflictsManager - deals with project level conflicts
and aggregation of the project rulesets so that the same rule set
will be analyzed/fixed only once.
2. Added RuleSetConflictController that exposes the conflicts to the user
and allowing to auto fix them.
3. There are various data classes that I created to abstract each layer.
4. Fixed a bug in the inspector that I found once integrated everything
(rules with None action would be considered missing).

Note: the tests for the output messages are low fidelity and focus on various
keywords that I'm expecting to find instead of the whole message (a nightmare to
maintain and you would still need to actually see the UX when changing them, so
no much point in having high fidelity tests for those).