SonarQube client (#651)
Refactoring SonarQubeService to simplify handling API changes in different versions of SonarQube

The service provides a set of logical APIs that are version-agnostic e.g. _GetProjectDashboardUrl_.
Internally, there are one or more version-dependent request classes wrap the calls to the server required to implement the logical API, using the server APIs that are available in particular version of SonarQube.

The service checks the version of SonarQube being accessed on initialisation. When a logical API request is made by a caller, the service chooses the appropriate version-dependent request class to handle the request.