SLVS-1487 Fix null reference exception in ManageBindingViewModel.CreteConnectionInfoFromSharedBinding by making sure that the shared binding configuration is always read even when the IsCurrentProjectBound is true.
It can be the case that, when the SuggestSharedBindingGoldBar detects the shared binding config file, the project is no yet bound. But when the window is opened by using the "Bind" button from the goldbar, the project is already bound. In this case there is no reason why the shared binding config should not be read.