SLVS-1710 Fix clang-cl compiler path
we add a new fallback way to get the compiler path correctly.
This fixes the `clang-cl` support, which is currently broken because the
`ClCompilerPath` property we rely on evaluates to `/clang-cl.exe`.
The new fallback method searches in the `ExecutablePath` path list for
the `ClToolExe` executable, which gives the correct answer in the case
of `clang-cl.exe`

Co-authored-by: Georgii Borovinskikh <georgii.borovinskikh@sonarsource.com>
Co-authored-by: Michael Jabbour <117195239+michael-jabbour-sonarsource@users.noreply.github.com>