SLVS-1856 Clean up non-roslyn quality profiles (#6060)
[SLVS-1856](https://sonarsource.atlassian.net/browse/SLVS-1856)

Upgrading to the new version will not clean the `binding.config` to
maintain backwards compatibility. The `binding.config` is not
regenerated and in case one of missing entries, it will just not sync
the missing entries. It will not break the Extension but there is a
chance that missing languages from the `binding.config` will not be
synchronized.

In this case a user can
* Update all instances of SQVS to the latest version.
* Unbind and re-bind the project to have a clean `binding.config`.

New bindings will only populate C# and VB.NET on the `binding.config`.

[SLVS-1856]:
https://sonarsource.atlassian.net/browse/SLVS-1856?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ